Add CheckNeighborVisibility; Create  new class BlockSource
Change the section 3D array into a dictionary  from BlockSource

+++ Assets/Scripts/Classes/Block.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Block
 {
-    public int Id;
+    public short Id;
     public string Name;
     public Vector3Int Position;// The absolute position in the map
     public GameObject BlockObject; // If the block is not created, this variable will be null
